Reorganizes as module (implements #2)
