+++ ffbot/__init__.py
@@ -1,0 +1,5 @@
+from .constants import VERSION
+from .optimizer import optimize
+from .scraper import scrape
+
+__version__ = VERSION
