+++ ffbot/constants.py
@@ -1,0 +1,1 @@
+VERSION = '1.0.0'
